Record players via ID
As part of building out our DB seeding, we need to be able to create
Players. Thus, we introduce the Player model and DB table. Of course
this is not all of the information that we need for a Player, but it
is a good start.

This will eventually be hooked into the ability to populate all data
for a team so it remains unused until then.