Create links between models
This is a large commit, but it does a lot of fleshing out of ideas.
These ideas are the relationships between the models which I will describe
here:
- A Player is any player, active or not
- A Team is any team, active or not
- A Roster is a snapshot of all Players on a Team at any given time
- A RosterAssignment is the relationship between a Player and a Roster
- A Player can have active and non-active RosterAssignments
- A Roster can be active or non-active
- A Player can only have one active RosterAssignment
- A Team can only have one active Roster

Through these rules, we can seed our database such that we:
- Run PullTeams::record_teams first
- Run our PullRosters::record_rosters which will in turn run PullPlayers.record_players
and create a RosterAssignment for them.

This will all be added to a PopulateData so that this won't need to be
remembered as for now the order does matter.