Create Seasons and Events
When doing some initial modelling for seasons, it occurred to me that
we will also need to display events such as the Olympics. Both of these
types of TimePeriod will need to be associated to player stats so these
will eventually change a bit, but this will start us off.