Create populate data method
Now that things are getting a little more in shape, we want to have
the ability to run one single script in order to populate our data.
This will now handle the fetching and persistence of all data.

This has been manually tested with the following command:
NhlApi::PopulateData.new.populate(start_year: 2020, end_year: 2021)