Fix up Roster logic
Unfortunately, this isn't a clean fix as of yet, but it fixes an issue
where we were always creating a new Roster/RosterAssignment and marking
it as active and everything else as inactive.

This temporary way just checks the year against the current year "20232024"
and does things differently for the most recent rosters.