Allow for pulling stats
This new class—PullStats—isn't very nuanced, but it's a great start.
It seems as though we need to pull a player's stats through their
landing page and iterate through the seasons in order to build our
stats.

This currently works for NHL stats so long as the player has a
RosterAssignment for the given season. Unfortunately it does not attribute
regular season and playoff stats yet so that will have to be added.