Display a player's seasons
By calculating the min and max years for a user, we can display the
length of their career. This is highly based off available data so
if your dataset is incomplete you will only get what your data says.

For accurate data, you would need to populate the entirety of all data.