Introduce null object pattern
Rather than have & strewn all around our view partials, it would be
much cleaner to introduce the null object pattern for stats so that
these can return nil on method_missing and be coerced into 0 for that
stat. This way we can still show a page that is erroring but have it
be silently erroring. Errors like this should be caught in specs first
though.