Separate forms and results
This change is a move toward having separate forms for the Query Builder
based on what the user chooses. If they chose "All Seasons", then we
ransack on the Player model and get our aggregate stats. If they chose
a specific season, then we query on the PlayerStat model and build our
ransackable attributes that way.

Currently, the results use different tables, but this will be reconciled.